Refactored to a separate js_notifications.hpp
