+++ src/js_realm_object.hpp
@@ -33,6 +33,7 @@ struct RealmObjectClass;
 #include "js_util.hpp"
 #include "js_realm.hpp"
 #include "js_schema.hpp"
+#include "js_notifications.hpp"
 
 namespace realm {
 namespace js {
@@ -41,96 +42,17 @@ template <typename>
 class NativeAccessor;
 
 template <typename T>
-class NotificationHandle;
-
-template <typename T>
-struct NotificationBucket {
-    using IdType = unsigned long long;
-    using ProtectedFunction = Protected<typename T::Function>;
-    using NotificationHandle = NotificationHandle<T>;
-    
-    static inline std::map<IdType, std::vector<std::pair<ProtectedFunction, NotificationToken>>> s_tokens;
-
-    static void emplace(NotificationHandle &handle, ProtectedFunction &&callback, NotificationToken &&token) {
-        if (handle) {
-            s_tokens[handle].emplace_back(std::move(callback), std::move(token));
-        } else {
-            throw std::runtime_error("Cannot emplace notifications using an unset handle");
-        }
-    }
-    
-    static void clear() {
-        s_tokens.clear();
-    }
-    
-    static void erase(NotificationHandle &handle) {
-        if (handle) {
-            s_tokens.erase(handle);
-        }
-    }
-    
-    static void erase(NotificationHandle &handle, ProtectedFunction &&callback) {
-        if (handle) {
-            auto& tokens = s_tokens[handle];
-            auto compare = [&callback](auto&& token) {
-                return typename ProtectedFunction::Comparator()(token.first, callback);
-            };
-            tokens.erase(std::remove_if(tokens.begin(), tokens.end(), compare), tokens.end());
-        } else {
-            throw std::runtime_error("Cannot erase notifications using an unset handle");
-        }
-    }
-};
-
-template <typename T>
-class NotificationHandle {
-    using NotificationBucket = NotificationBucket<T>;
-    using IdType = typename NotificationBucket::IdType;
-    
-    static inline IdType s_next_id = 0;
-    std::optional<IdType> m_id;
-
-public:
-    NotificationHandle() {
-        m_id = s_next_id;
-        if (s_next_id == std::numeric_limits<IdType>::max()) {
-            throw std::overflow_error("No more NotificationHandle ids");
-        }
-        s_next_id += 1;
-    };
-    
-    ~NotificationHandle() {
-        NotificationBucket::erase(*this);
-    }
-    
-    operator IdType() const {
-        return *m_id;
-    };
-    
-    operator bool() const {
-        return m_id.has_value();
-    };
-    
-    NotificationHandle(NotificationHandle &&other) {
-        m_id = std::move(other.m_id);
-        other.m_id.reset();
-    }
-};
-
-
-
-template <typename T>
 class RealmObject : public realm::Object {
 public:
     RealmObject(RealmObject const& obj)
         : realm::Object(obj){};
     RealmObject(realm::Object const& obj)
         : realm::Object(obj){};
-    RealmObject(RealmObject &&other) = default;
+    RealmObject(RealmObject&& other) = default;
     RealmObject& operator=(RealmObject&&) = default;
     RealmObject& operator=(RealmObject const&) = default;
-    
-    NotificationHandle<T> m_notification_handle;
+
+    notifications::NotificationHandle<T> m_notification_handle;
 };
 
 template <typename T>
@@ -145,7 +67,7 @@ struct RealmObjectClass : ClassDefinition<T, realm::js::RealmObject<T>> {
     using Function = js::Function<T>;
     using ReturnValue = js::ReturnValue<T>;
     using Arguments = js::Arguments<T>;
-    using NotificationBucket = NotificationBucket<T>;
+    using NotificationBucket = notifications::NotificationBucket<T>;
 
     static ObjectType create_instance(ContextType, realm::js::RealmObject<T>);
 
@@ -483,7 +405,6 @@ void RealmObjectClass<T>::add_listener(ContextType ctx, ObjectType this_object, 
         Function::callback(protected_ctx, protected_callback, protected_this, 2, arguments);
     });
     NotificationBucket::emplace(realm_object->m_notification_handle, std::move(protected_callback), std::move(token));
-    //realm_object->m_notification_tokens.emplace_back(protected_callback, std::move(token));
 }
 
 template <typename T>
@@ -500,13 +421,6 @@ void RealmObjectClass<T>::remove_listener(ContextType ctx, ObjectType this_objec
         throw std::runtime_error("Invalid 'this' object");
     }
 
-    /*
-    auto& tokens = realm_object->m_notification_tokens;
-    auto compare = [&](auto&& token) {
-        return typename Protected<FunctionType>::Comparator()(token.first, protected_function);
-    };
-    tokens.erase(std::remove_if(tokens.begin(), tokens.end(), compare), tokens.end());
-    */
     NotificationBucket::erase(realm_object->m_notification_handle, std::move(protected_callback));
 }
 
@@ -520,9 +434,8 @@ void RealmObjectClass<T>::remove_all_listeners(ContextType ctx, ObjectType this_
     if (!realm_object) {
         throw std::runtime_error("Invalid 'this' object");
     }
-    
+
     NotificationBucket::erase(realm_object->m_notification_handle);
-    // realm_object->m_notification_tokens.clear();
 }
 
 template <typename T>
